requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
yfinance==0.2.28
plotly==5.18.0
pandas==2.1.4
gunicorn==21.2.0